moving popup related classes into a Popups folder
