+++ XLPrecisionKeyframes/UserInterface/Popups/UserInterfacePopup.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 using XLPrecisionKeyframes.Keyframes;
 
-namespace XLPrecisionKeyframes.UserInterface
+namespace XLPrecisionKeyframes.UserInterface.Popups
 {
     public class UserInterfacePopup<T> where T : EditBaseUI
     {
